fix SceneView.drawGizmos bug (see https://github.com/radiatoryang/hedera/issues/9 ), change LightmapStatic to ContributeGI for 2019.2+, moved Examples to Samples folder (for UPM), update package.json to reflect new package
