+++ Editor/Scripts/IvyMesh.cs
@@ -50,7 +50,11 @@ namespace Hedera
                 ivyGraph.rootGO.transform.SetParent( ivyGraph.rootBehavior );
             }
             SetStaticEditorFlag( ivyGraph.rootGO, StaticEditorFlags.BatchingStatic, ivyProfile.markMeshAsStatic );
+            #if UNITY_2019_2_OR_NEWER
+            SetStaticEditorFlag( ivyGraph.rootGO, StaticEditorFlags.ContributeGI, ivyProfile.useLightmapping );
+            #else
             SetStaticEditorFlag( ivyGraph.rootGO, StaticEditorFlags.LightmapStatic, ivyProfile.useLightmapping );
+            #endif
             var rootTrans = ivyGraph.rootGO.transform;
             rootTrans.position = ivyGraph.seedPos;
             rootTrans.rotation = Quaternion.identity;
@@ -186,7 +190,11 @@ namespace Hedera
 
         static void RefreshMeshObject(MeshFilter mf, IvyProfile ivyProfile) {
             SetStaticEditorFlag( mf.gameObject, StaticEditorFlags.BatchingStatic, ivyProfile.markMeshAsStatic );
+            #if UNITY_2019_2_OR_NEWER
+            SetStaticEditorFlag( mf.gameObject, StaticEditorFlags.ContributeGI, ivyProfile.useLightmapping );
+            #else
             SetStaticEditorFlag( mf.gameObject, StaticEditorFlags.LightmapStatic, ivyProfile.useLightmapping );
+            #endif
             var branchTrans = mf.transform;
             branchTrans.localPosition = Vector3.zero;
             branchTrans.localRotation = Quaternion.identity;
