+++ Samples/ExampleIvyScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65ac718cee8d2294eafbc08d78479cfb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
