+++ Samples/ExampleIvyScene/ExampleIvyMeshExport.obj.meta
