+++ Samples/ExampleIvyScene/HederaExampleScene/HederaData.asset.meta
