+++ Samples/ExampleIvyScene/HederaExampleScene/LightingData.asset
