+++ Samples/ExampleIvyScene/HederaExampleScene/LightingData.asset.meta
