+++ Samples/ExampleIvyScene/HederaExampleScene/ReflectionProbe-0.exr.meta
