+++ Samples/ExampleIvyScene/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eaf96272705dc464b8c33795a02727b5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
