+++ Samples/ExampleIvyScene/Scripts/OpenHederaExampleScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f0800537eadde064abaccc7f3e1c4acd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
