+++ package.json
@@ -2,8 +2,8 @@
 "name": "com.radiatoryang.hedera",
 "displayName": "Hedera",
 "description": "Paint 3D ivy in the editor and simulate ivy growth patterns in your scene.",
-"version": "1.0.0",
-"unity": "2018.3",
+"version": "1.2.0",
+"unity": "2019.4",
 "license": "GPL 2.0",
 "dependencies": {
  }
